+++ packages/zeebe-element-templates-json-schema/test/fixtures/invalid-binding-type.js
@@ -47,14 +47,15 @@ export const errors = [
               'bpmn:Message#property',
               'bpmn:Message#zeebe:subscription#property',
               'zeebe:taskDefinition',
-              'zeebe:calledElement'
+              'zeebe:calledElement',
+              'zeebe:linkedResource'
             ]
           },
           message: 'should be equal to one of the allowed values'
         }
       ]
     },
-    message: 'invalid property.binding type "foo"; must be any of { property, zeebe:taskDefinition:type, zeebe:input, zeebe:output, zeebe:property, zeebe:taskHeader, bpmn:Message#property, bpmn:Message#zeebe:subscription#property }'
+    message: 'invalid property.binding type "foo"; must be any of { property, zeebe:taskDefinition:type, zeebe:input, zeebe:output, zeebe:property, zeebe:taskHeader, bpmn:Message#property, bpmn:Message#zeebe:subscription#property, zeebe:taskDefinition, zeebe:calledElement, zeebe:linkedResource }'
   },
   {
     dataPath: '',