+++ packages/zeebe-element-templates-json-schema/CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to [@camunda/zeebe-element-templates-json-schema](https://gi
 
 ___Note:__ Yet to be released changes appear here._
 
+* `FEAT`: support `zeebe:linkedResource` binding
+
 ## 0.21.0
 
 * `FEAT`: support `engines` field ([#146](https://github.com/camunda/element-templates-json-schema/issues/146))